Fix: Rename monitoring service file
Renamed `src/services/monitoring_service.ts` to `src/services/monitoring-service.ts` to follow the preferred naming convention.

The import path in `src/commands/default.ts` and a JSDoc reference to the filename have been updated accordingly.
